+++ tests/utils/test_utils.py
@@ -33,8 +33,6 @@ from BALSAMIC.utils.cli import (
     check_executable,
     job_id_dump_to_yaml,
     generate_h5,
-    get_md5,
-    create_md5,
     get_sample_dict,
     get_pon_sample_dict,
     get_fastq_files_directory,
@@ -43,7 +41,6 @@ from BALSAMIC.utils.cli import (
 from BALSAMIC.utils.io import read_json, write_json, read_yaml
 
 from BALSAMIC.utils.rule import (
-    get_chrom,
     get_vcf,
     get_sample_id_by_type,
     get_picard_mrkdup,
@@ -52,7 +49,6 @@ from BALSAMIC.utils.rule import (
     get_result_dir,
     get_threads,
     get_delivery_id,
-    get_reference_output_files,
     get_rule_output,
     get_sample_type_from_prefix,
 )
@@ -159,18 +155,6 @@ def test_get_variant_callers_wrong_sequencing_type(tumor_normal_config):
             mutation_class=mutation_class,
             sequencing_type=wrong_sequencing_type,
         )
-
-
-def test_get_reference_output_files():
-    # GIVEN a reference genome version
-    genome_ver = "hg38"
-    file_type = "fasta"
-
-    # WHEN getting list of valid types
-    fasta_files = get_reference_output_files(REFERENCE_FILES[genome_ver], file_type)
-
-    # THEN it should return list of file
-    assert "Homo_sapiens_assembly38.fasta" in fasta_files
 
 
 def test_get_bioinfo_tools_version():
@@ -354,7 +338,7 @@ def test_get_snakefile():
             elif analysis_type == "generate_ref" and reference_genome != "canfam3":
                 pipeline = "BALSAMIC/workflows/reference.smk"
             elif analysis_type == "generate_ref" and reference_genome == "canfam3":
-                pipeline = "BALSAMIC/workflows/reference-canfam3.smk"
+                pipeline = "BALSAMIC/workflows/reference_canfam3.smk"
             elif analysis_type == "pon":
                 pipeline = "BALSAMIC/workflows/PON.smk"
             elif analysis_workflow == "balsamic-qc":
@@ -365,31 +349,6 @@ def test_get_snakefile():
             assert snakefile.startswith("/")
             assert pipeline in snakefile
             assert Path(snakefile).is_file()
-
-
-def test_get_chrom(panel_bed_file: str):
-    # Given a panel bed file
-    actual_chrom = [
-        "10",
-        "11",
-        "16",
-        "17",
-        "18",
-        "19",
-        "2",
-        "3",
-        "4",
-        "6",
-        "7",
-        "9",
-        "X",
-    ]
-
-    # WHEN passing this bed file
-    test_chrom = get_chrom(panel_bed_file)
-
-    # THEN It should return list of chrom presents in that bed file
-    assert set(actual_chrom) == set(test_chrom)
 
 
 def test_get_vcf(sample_config):
@@ -815,41 +774,6 @@ def test_generate_h5_capture_no_output(tmp_path):
         actual_output = generate_h5(dummy_job_name, dummy_job_id, dummy_path)
 
     assert actual_output == None
-
-
-def test_get_md5(tmp_path):
-
-    # GIVEN a dummy file
-    dummy_dir = tmp_path / "md5"
-    dummy_dir.mkdir()
-    dummy_file = dummy_dir / "dummy_file.dump"
-    dummy_file.write_text("Awesome Text")
-
-    # THEN md5 returned should be
-    assert get_md5(dummy_file) == "3945B39E"
-
-
-def test_create_md5(tmp_path):
-
-    # GIVEN a path to a md5 file and reference dummy files
-    ref_dir = tmp_path / "references"
-    ref_dir.mkdir()
-    dummy_ref_file1 = ref_dir / "reference_file1.dump"
-    dummy_ref_file1.write_text("Test reference1")
-    dummy_ref_file2 = ref_dir / "reference_file2.dump"
-    dummy_ref_file2.write_text("Test reference2")
-    dummy_reference_dict = {
-        "reference_dummy1": str(dummy_ref_file1),
-        "reference_dummy2": str(dummy_ref_file2),
-    }
-    dummy_dir = tmp_path / "md5"
-    dummy_dir.mkdir()
-    dummy_file = dummy_dir / "dummy_file.dump"
-
-    create_md5(dummy_reference_dict, dummy_file)
-
-    # THEN md5 file exists
-    assert dummy_file.exists()
 
 
 def test_get_rule_output(snakemake_fastqc_rule):