+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changed:
 * Updated cryptography version to 41.0.1 https://github.com/Clinical-Genomics/BALSAMIC/pull/1173
 * Refactor constants https://github.com/Clinical-Genomics/BALSAMIC/pull/1174
 * Move models to their own folder https://github.com/Clinical-Genomics/BALSAMIC/pull/1176
+* Balsamic init workflow refactoring https://github.com/Clinical-Genomics/BALSAMIC/pull/1188
 
 Fixed:
 ^^^^^^